requirements.txt
streamlit
python-dotenv
requests
beautifulsoup4
playwright
google-generativeai
faiss-cpu
langchain-google-genai
langchain
langchain-community
tenacity
